Make snake head update on direction change
Previously snake parts were static until a part became the tail at which
point it was re-orientated if necessary to match the direction of the
preceding part. The new approach is simpler and is consistent for the
whole snake, so that any direction-specific design of the snake in the
UI will display correctly.